requirements/base.txt
python-slugify==8.0.4  # https://github.com/un33k/python-slugify
Pillow==10.2.0  # https://github.com/python-pillow/Pillow
argon2-cffi==23.1.0  # https://github.com/hynek/argon2_cffi
whitenoise==6.6.0  # https://github.com/evansd/whitenoise
redis==5.0.3  # https://github.com/redis/redis-py
hiredis==2.3.2  # https://github.com/redis/hiredis-py

# Django
# ------------------------------------------------------------------------------
django==4.2.11  # pyup: < 5.0  # https://www.djangoproject.com/
django-environ==0.11.2  # https://github.com/joke2k/django-environ
django-model-utils==4.4.0  # https://github.com/jazzband/django-model-utils
#django-allauth[mfa]==0.61.1  # https://github.com/pennersr/django-allauth
django-allauth[mfa] @ git+https://github.com/347S24/django-allauth.git
# -e /Users/tgm/dev/dj-aa/
django-crispy-forms==2.1  # https://github.com/django-crispy-forms/django-crispy-forms
crispy-bootstrap5==2024.2  # https://github.com/django-crispy-forms/crispy-bootstrap5
django-redis==5.4.0  # https://github.com/jazzband/django-redis
django-ninja==1.1.0

# Canvas
canvasapi==3.2.0

requests==2.31.0